[#170] test : token, userStatus 저장
