+++ src/pages/signup/SignUp.tsx
@@ -16,74 +16,18 @@ function SignUp() {
   const [jwtToken, setJwtToken] = useState('')
   const [userStatus, setUserStatus] = useState('')
 
-  const setToken = (token: string) => {
-    setJwtToken(token)
-    alert(`token ${token}`)
-    console.log('token', token)
-  }
-  const setStatus = (status: string) => {
-    setUserStatus(status)
-    alert(`status ${status}`)
-    console.log('status', status)
-    storage.set('userStatus', status)
-  }
-
-  // Native to JS로 토큰, status 저장
   useEffect(() => {
-    window.setToken = setToken as (token: string) => void
-    window.setUserStatus = setStatus as (status: string) => void
-  }, [])
-
-  useEffect(() => {
-    if (jwtToken) {
-      storage.set('accessToken', jwtToken)
-      alert(`NATIVE to REACT jwtToken 설정 완료 ${jwtToken}`)
-    } else {
-      alert('jwtToken 없음')
+    const token = storage.get('accessToken')
+    const status = storage.get('userStatus')
+    if (token) {
+      console.log(token)
+      alert(`token : ${token}`)
+      setJwtToken(token)
     }
-  }, [jwtToken])
-
-  useEffect(() => {
-    const handleTokenEvent = (event: MessageEvent<string>) => {
-      if (event.data) {
-        try {
-          const tokenData = JSON.parse(event.data)
-          if (tokenData.token) {
-            storage.set('accessToken', tokenData.token)
-            setToken(tokenData.token)
-            alert(`Received token: ${tokenData.token}`)
-          }
-        } catch (error) {
-          console.error('Error parsing token data:', error)
-          alert('error')
-          // Handle parsing error appropriately (e.g., display an error message to the user)
-        }
-      }
-    }
-
-    const handleStatusEvent = (event: MessageEvent<string>) => {
-      if (event.data) {
-        try {
-          const statusData = JSON.parse(event.data)
-          if (statusData.status) {
-            setStatus(statusData.status)
-            storage.set('userStatus', statusData.status)
-            alert(`Received user status: ${statusData.status}`)
-          }
-        } catch (error) {
-          console.error('Error parsing status data:', error)
-          // Handle parsing error appropriately
-          alert('error')
-        }
-      }
-    }
-
-    window.addEventListener('setToken', handleTokenEvent)
-    window.addEventListener('setUserStatus', handleStatusEvent)
-
-    return () => {
-      window.removeEventListener('setToken', handleTokenEvent)
-      window.removeEventListener('setUserStatus', handleStatusEvent)
+    if (status) {
+      console.log(status)
+      alert(`status : ${status}`)
+      setUserStatus(status)
     }
   }, [])
